Fix PG hearts_balance: cast params in addition (operator not unique unknown+unknown)
Made-with: Cursor

+++ services/oneoffUserCleanupMarch2026.js
@@ -268,7 +268,7 @@ async function runMarch2026AunyaweePhongCleanup(options = {}) {
         `UPDATE users SET
           red_giveaway_balance = $2,
           red_hearts_balance = $3,
-          hearts_balance = $4 + $3 + $2
+          hearts_balance = $4::integer + $3::integer + $2::integer
          WHERE id = $1::uuid`,
         [auId, newGive, newRed, pink]
       );
